+++ extension/functions/src/ansi.rs
@@ -14,3 +14,72 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
+
+use arrow::array::{ArrayRef, Int64Array};
+use arrow::datatypes::DataType;
+use datafusion::error::Result;
+use datafusion::logical_expr::Volatility;
+use datafusion_common::cast::{as_string_array};
+use datafusion_expr::{
+    ReturnTypeFunction, ScalarFunctionDef, ScalarFunctionPackage, Signature,
+};
+use std::sync::Arc;
+
+#[derive(Debug)]
+pub struct LengthFunction;
+
+impl ScalarFunctionDef for LengthFunction {
+    fn name(&self) -> &str {
+        "length"
+    }
+
+    fn signature(&self) -> Signature {
+        Signature::exact(vec![DataType::Utf8], Volatility::Immutable)
+    }
+
+    fn return_type(&self) -> ReturnTypeFunction {
+        let return_type = Arc::new(DataType::Int64);
+        Arc::new(move |_| Ok(return_type.clone()))
+    }
+
+    fn execute(&self, args: &[ArrayRef]) -> Result<ArrayRef> {
+        assert_eq!(args.len(), 1);
+        let input = as_string_array(&args[0]).expect("cast failed");
+        let array = input
+            .iter()
+            .map(|value| match value {
+                Some(text) => Some(text.chars().count() as i64),
+                _ => None,
+            })
+            .collect::<Int64Array>();
+        Ok(Arc::new(array) as ArrayRef)
+    }
+}
+
+// Function package declaration
+pub struct FunctionPackage;
+
+impl ScalarFunctionPackage for FunctionPackage {
+    fn functions(&self) -> Vec<Box<dyn ScalarFunctionDef>> {
+        vec![Box::new(LengthFunction)]
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use datafusion::error::Result;
+    use datafusion::prelude::SessionContext;
+    use tokio;
+
+    use crate::utils::{execute, test_expression};
+
+    use super::FunctionPackage;
+
+    #[tokio::test]
+    async fn test_length() -> Result<()> {
+        test_expression!("length('jose')", "4");
+        test_expression!("length('oooppp?')", "7");
+        Ok(())
+    }
+    
+}
